draft 2
This is my most recent draft of my website. It's taking me longer than I initially thought. I'm planning on adding a picture to the background, pictures to the blank boxes, all the text content, and fix the alignment of the boxes. Once finished, this will be a portfolio to house my resume and creative work.